requirements.txt
pyspark
cassandra-driver
mysql-connector-python==8.0.29
